Reorged entity worker tests
